livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.61",
    "inngest>=0.5.18",
    "livekit-agents[aws,google,mcp]~=1.4",
    "livekit-plugins-aws[realtime]>=1.4.5",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
